Two programs to fix errors in tables
vtcopymissingrows will copy rows that are missing from the source. This sometimes happens when we shard split because we have uniqueness constraint violations and Vitess doesn't handle them very well. Sometimes we drop a few chunks on e.g. aliases and instrument_links.

vtrepairtable moves rows that for some reason are in the wrong shard to the correct shard.