Add tool vtcopytable to copy a table from a keyspace (unsharded) to another keyspace.
This first version will work if the table is blacklisted in the source keyspace as it does not track ongoing changes to the table that is going to be copied over. It also does just the table copying only. So blacklisting, copying schema over to detination shards and applying vschemas must be done before using this tool